Properly used GNU Make. (Part 3.)
I'll try to summarise my understanding of how GNU Make works.

A target of the form:
```
foo: bar
	baz
```
creates the output file `foo`. The files `bar` are considered inputs: they are created first; if they are unchanged from the last run and `foo` has a newer timestamp than them, `foo` is not re-generated using the `baz` command.

The files `bar` are processed recursively: to create them (which happens before `foo` is created, as described above), Make first checks their inputs: those inputs are created first, and if they are unchanged from the last run and `bar` have a newer timestamp than them, `bar` are not re-generated.