Removed unwanted `undef`s. Optimised sudoku generator.
Macros are visible only in the file where they are defined, so there is no need to `undef` them at the end of the file (unless they are declared in headers).

The sudoku generator previously generated indices coutinuously until the required number of deletions was achieved. Now, the indices are generated in such a way that there will be no repetitions, so the generator can simply delete the number without having to check if the number has already been deleted.