Attempted to add a reentrant variant of the PRNG.
And undid that attempt (I rebased the commits), because it would require exposing the internals of the `mt19937` structure to the user, which I did not want to do. (I wanted to be able to generate random numbers using an instance of the RNG, similar to the C++ MT19937, but also give the user the option to use the global-state non-reentrant RNG.) Further, I made the test simpler, but if it fails, the program will exit.

Removed all references to MT19937 in the main program. Let the sudoku solver deal with the PRNG. No need to expose those details to the main program.