Fixed C++ makefiles. Switched shell scripts to `sh`.
The C++ compiler is specified using the `CXX` and `CPPFLAGS` variables in GNU Make.

`sh` (the shell command language) is specified by the POSIX standard. It supports command substitution, so the shell scripts will work with plain `sh` as well. Implementations of it (e.g. `bash`) will be able to execute them properly. The syntax I have used isn't compatible with `csh` and `tcsh`, but they are not implementations of `sh`, so there should be no problems.

a