Added generation step immediately after twisting.
This allows the generation loop to be vectorised. (Generating in the twist loop itself results in poorer vectorisation.) Telling the compiler to unroll loops decreases the branch prediction accuracy from around 99% to 96%, but increases overall performance.

The only drawback of these changes is that the stack space required is doubled to around 10 KiB.