Added Python benchmarks.
I have not yet figured out why seeding in Python is as fast as in C++, but generation is a hundred times slower. My classmates suggested that Python adds a constant delay to all the times (which is why the ratio is almost 1 for seeding times but almost 100 for generation times); this seems correct, but I must confirm it.