Increased the number of iterations.
There is one more piece of the puzzle I am yet to figure out. On almost all computers I tested, for some reason, `mt19937.real32` is faster than `mt19937.rand32`. On a computer with a powerful CPU, this applies for the C functions as well: `mt19937_real32` is faster than `mt19937_rand32`, even though the former calls the latter. Mysterious. I increased the number of iterations (but the new number is still a multiple of 0x270 to avoid over-representing the twist delay) because that seemed to alleviate this issue in an MWE which reproduced this confounding benchmark. It did not alleviate the issue in this benchmark, though. I don't know how to solve this.