Updated comments and error messages.
Noted what the C standard says about the representations of integers, as a result of which, this code has well-defined effects (i.e. does not invoke implementation-defined behaviour). Python API error messages updated to hint about the rare case when the range of `int32_t`/`int64_t` is a proper superset of the range of `int long`/`int long long`.