Alternative constructors to non-deterministically seed.
In C++, the user had to create an MT19937 object (which would imply default seeding during construction) and then call its `init32` or `init64` method (which would seed again). The new constructors call `init32` or `init64` directly, avoiding repeated seeding. They differ from the other constructors in that they take a `char const*` instead of an unsigned integer, so the user can pass `""` as the argument. (I could have made them take `void*`, but then the user would have to pass `NULL` as the argument, which is not guaranteed to be a pointer. The C++ standard allows it to be 0. I could also have made them take `nullptr_t`, but it supposed to be available in C23.)