Do not include C headers in C++.
`mt19937.h` should include `cinttypes` and `cstddef` in C++. As a result, the exact-width integer types have to be `#defined`d (because they may not be in the global namespace) so that the same code can work in C, and `#undef`d at the end. Using the C++ header allows the new constructors introduced in 36fce247901efdd28bd6c448a1bd7cc865488586 to take `nullptr` as the first argument, since `nullptr` was introduced in C++11.