Changed the C++ benchmarking program back to use a macro.
It was written using a macro earlier, but I changed it to use a function template at some point. However, the C++ functions cannot be benchmarked then, because their types are not completely known at the time of template expansion (because they are themselves defined using variadic templates). (After all, the C++ functions are more intuitive to use, so _they_ should be benchmarked, although they will most probably get replaced with the C functions as part of compiler optimisations, because they are trivial wrappers around the latter.) Using a macro allows this.